Each exporter client is now a separate thread
This fixes problems caused by a single client asking for a lot of history and
essentially blocking anyone else from connecting to NNTSC and / or requesting
data.

It seems to be functional in the general case, e.g. requesting aggregate data,
collections, live data -- at worst, I've not broken anything that used to work.
Testing the export of new streams will have to be done with the "production"
setup, most likely when we start using AMP more.

Will test whether the change has any benefits soon -- again, we need production
NNTSC with lots of data to test this properly.

Also, fixed the rrdtool dependency in setup.py as we were requiring the wrong
rrdtool python library.