LPI server disconnect is no longer fatal
If the connection to the LPI server fails or breaks for some reason, we will
now attempt to reconnect (with some exponential back-off) rather than just
have the LPI process die.

This is great for cases where I have made a change to the lpi collector and
need to restart it. Previously, this would require me to also restart NNTSC
and any long-running clients attached to it, e.g. netevmon. Now I can restart
the collector and NNTSC will eventually reconnect and resume data collection.