Return a useful binsize for finding missing data
The 'binsize' field in the NNTSC_HISTORY message is now used to tell the
client what the measurement frequency should be, given the binsize that they
requested. The client can then use this information to determine if there
are missing measurements -- use case: inserting breaks into line graphs where
data is missing rather than drawing a straight line between the points either
side of the missing data.

In essence, it works as follows:
	if the requested binsize is greater than the measurement frequency:
		return requested binsize
	else:
		return measurement frequency

There are a bunch of heuristics inside lib/database.py now for determining
which of the above options is the case and what the measurement frequency is.
These have been documented extensively.

Each data point returned by the database now has a 'binstart' field which
contains the timestamp marking the start of the bin covered by that data.
This was previously named 'timestamp' and was getting overwritten by another
field named 'timestamp'.