Added dataparsers for remaining LPI metrics
Added dataparsers for LPI packets, LPI flows (both peak and new) and LPI users
(both active and observed). This covers all the metrics supported by the
current version of the LPI collector.

Changed the naming scheme for LPI bytes streams -- basically we use the word
"outgoing" rather than "out" and "incoming" instead of "in" so the titles are
less of an abomination against the English language.

Changed the internal names for the metrics in lpi_common, just to get rid of
the ugly underscores and make sure everything is consistent.