LPI replaces missing values with zeroes
The LPI collector no longer exports counters for protocols that have the value
of zero. This saves bandwidth and processing time, but means that our client
needs to determine which protocols are missing from the received data so
that zero values are pushed through to the data parsers for those protocols.

We also no longer create streams if we have not observed a non-zero value
for a particular time series. This saves us a lot of streams once you start
doing per-user data collection -- individual users won't use most protocols
so there's not much point in filling our database up with a time series
containing only zero.