Deal with the case where a selector column is listed multiple times.
It needs to iterate over the list of selectors and match them with columns, rather than iterating over the list of columns and matching them with selectors. This would be used in the case where multiple different aggregation functions should be applied to the same column (for example calculating average and standard deviation of a column).