Add ability to fetch AMP ICMP data formatted for smokeping graphs.
Using highly binned AMP data we can extract various percentiles (really "vigintiles" in this case as there are 20 of them) to use in a similar fashion to the 20 ping values smokeping collects. The database splits the sorted RTT values within a bin into 20 groups, and takes the average value across each group, presenting these as an array that can be used to plot smoke values on a graph.

Currently it's rather specific to the AMP ICMP case, but should be able to be made more generic.