Fix bug with duplicate data insertions
Basically, we can't rely on a return value from inserting into a partitioned
table as we need to return NULL to prevent the double insert.

Changed approach for getting the test id after inserting a new traceroute
test. It is a bit hax, but at least works with partitions.

Disabled implicit returns for our database engine. This fixes the crashes that
occur if a partitioned table had a primary key.