AMP ICMP streams now include the address in the stream definition
This will allow us to more easily compare and display the results for tests
to a particular address returned by DNS for the destination. Some targets
have multiple addresses (a notable example being having a v4 and v6 address)
and other targets will change addresses from test to test due to load balancing
etc. -- now we will be able to treat each address as a separate stream.

I've also added some code to insert null data values for streams that do not
report a measurement result, e.g. because that address was not tested to that
time around. We probably don't need or want this behaviour, but there is an
outside chance of things that rely on the measurements having a consistent
frequency breaking now that we have AMP ICMP streams that may only be
measured intermittently (netevmon, ampy determining where gaps in the data
are). If we do decide we need it, we can enable the code. If not, get rid
of it later.