Added first pass at some database code using psycopg for queries
Seems to work reasonably well in standalone tests, so the next job is to
get the exporter code using the DBSelector class instead of the Database
class.

Most of the API is the same as in Database, with a few exceptions:
 * select_percentile_data -- major changes to the parameters
 * the frequency returned by select_*_data is now a dictionary instead
   of a single value. The key is the stream id and the value is the
   frequency for that particular stream.

NOTE: this code won't work too well with amp-traceroute data yet. Still
need to implement the parameterised function needed for that collection.

Removed modules from lib/__init__.py that were no longer in the libnntsc
package.