Changed sanitise_columns to query information_schema rather than the data table
This was slow for complicated "tables" such as the select_amp_traceroute
procedure, because of the effort required just to get one row. Instead, we
extract the columns from the data table / view named for the collection
itself.

This does mean that we will need to keep the data_amp_traceroute view even
though we won't be querying it directly.