Fixed blocking issue when replacing the partition trigger
Dropping the trigger requires an exclusive lock on the table
the trigger belongs to, but under heavy query load (e.g. when starting
netevmon) this becomes a problem. Our partition code was blocking on
this and halting all further database inserts until the client went
away.

Now, we only create the trigger if it does not exist (we check for
existence by consulting metadata tables) and this creation only
happens on program start-up, before any clients can connect.