Database queries by the exporter are now made using worker threads
This fixes the problem where the NNTSCClient ignored live data and new
client messages because it was busy waiting for a SELECT query to
complete. Now these queries are farmed out to a set of worker
threads who run the query and feed the results back to the NNTSCClient
when done.

To avoid excessive memory consumption, large historical queries are
now broken down into smaller queries covering weekly intervals.

Also, frequency calculations and data formatting is now done inside
the DBSelector code, which allows us to avoid having to store all of
this information in memory. Instead, we can process each stream in
turn rather than trying to do them all at once.