Slightly nicer solution to the missing trigger function problem.
The main reason that this is nicer is because it doesn't cause the partition
code to think that it has seen a datapoint at "now" when it really hasn't.
This could have some bad side-effects down the line, so I think it is better
to be more explicit about what we are doing (even if it involves creating a
dummy trigger function).