Implement data collection for the AMP http test
Made a few modifications to the original table structure which may need to
be accounted for if updating an existing NNTSC install. Easiest way to fix
would be to manually drop the amp-http collection (there should be no data to
lose) and any associated tables, i.e. ones with http2 in the name.

Note the name change from http2 to http.