When selecting to find the row we just added, escape bad chars.
Our inserts were fine, but there are cases with complicated
multi-table collections (AMP traceroute and http) where we would insert a
row and then immediately try to query for that row to get the id created
for that row. These queries were simply formed using text() and
standard python variable substitution.

This was fine up until someone decided to add an object to their website
with a ' in the path, which broke these queries.

Of course, we should have been properly escaping these parameters regardless
so better to find and fix it now than get haxxed later on.