Add support for a new NNTSC_PUSH message to the NNTSC protocol
The PUSH message will be used to indicate to clients that the current block
of data received from a collector has been completely processed.

This is mainly useful for supporting grouping streams in anomalyfeed, where a
PUSH will indicate that there will be no more data for that stream group
until the next measurement interval and therefore anomalyfeed can export the
aggregated results immediately rather than waiting for a new measurement with
a different timestamp.