Parameterise label names and binsize in dbselect code
Removed code to convert stream lists to label dictionaries. This should now
happen inside nntscclient, which will ensure that all NNTSC code can rely on
working with label dicts only. Left some assertions in that require the labels
to be a dictionary -- this is mainly so we can catch any outstanding cases
where stream lists get through to dbselect but we should have got them all.