Wrapped all basic cursor operations with try/except
This will prevent problems where non-data queries would never
complete due to a timeout.

The current behaviour is simply to return nothing when this happens,
which may not be ideal. It would be nice if we could let the querier
know that their query timed out as opposed to there actually being
no streams / collections / etc.