Replaced pipes inside exporter.py with queues
ClientAPI changes:
 * increased the buffer size for recv() calls from 4k to 256k
 * Added a new message type: REGISTER_COLLECTION (used internally)

NNTSC messages are now formed by the worker threads and pushed onto a queue
for sending, rather than the full result of the query being pushed back to the
client thread and the client having to break it up into suitable messages.

Subscribe stream messages are now pushed onto the queue after all the history
has been pushed on. This streamlines the subscription process compared with
what it was previously (subend variables etc.)

Client threads have an "outstanding" member, which is used to store the
portion of a message that isn't sent when a non-blocking transmit finished
before sending the whole message.

Incremented version number to 2.4