Replaced exporter pipes with a rabbit queue
This is aimed at preventing the possibility of an exporter pipe
filling up and the dataparsers grinding to a halt due to being unable
to complete a blocking write.

Replaced the amp dataparser consumer with the consumer from our new
pikaqueue module.

The exporter now listens for new client connections in a separate thread.

Added locks around the subscriber maps within the exporter to protect them
from race conditions if multiple clients try to register new subscriptions
at the same time.

Removed the per-client "store" and "timer" variables as these are no longer
necessary since the danger of the live source filling up has more or less
disappeared.