Fix bug where live data would get interleaved with history
This would cause a crash due to insecure pickling.

The problem occurs when the history is not successfully sent in its
entirely, leaving some outstanding. Due to the way we were prioritising
live data, we would then move straight on to the next live segment rather
than sending the remainder of the half-sent history. The client, of course,
just gets a mess of interleaved messages that it can't decode properly and
goes WTF.