Ensure no measurements are missing when subscribing to streams
The main aim is to avoid missing data that arrives between the time
the queries for history are being made and the time that the queries
are completed.

Now we store any live measurements that arrive for streams of interest
while the queries are still underway. Once the queries complete
(signalled by a new HISTORY_DONE message), the saved measurements are
then sent to the client as though they were "live".

There are still some problems with querying for data that has been
processed but not committed to the database yet -- since it has been
processed it won't come through as live data that we can store, but
it also isn't in the database so we don't pick it up in the query.
This tends to happen when the history queried is quite short, i.e.
can be dealt with using a single query.