Add delay to final query to allow for recent data to be committed
This is a horrible hack to fix the problem where the most recent data
is not yet committed when we ask for history up until now. This often
meant we would have a small gap between the end of our history and the
start of our live data.

Having a big enforced delay in our query isn't a particularly nice
solution but it's not clear that there are any better options.