Moved insert code into database.py
All of the dataparsers had very similar code for inserting
new streams and data, so I've implemented generic insert functions
in database.py for the dataparsers to use instead.

Updated the dataparsers to use the new functions, with the
exception of amp_http as this is much more complicated.

Added a new return code for database inserts -- DB_INTERRUPTED --
which we will return if we get a KeyboardInterrupt during an insert
operation.