Added a timestamp to the HISTORY_DONE message
Now we collect all live data received after the subscribe began,
but use the timestamp to discard any live data that had already made
it into the database by the time we get around to querying the
historical data.

By doing this, we might finally be able to have an adequate
solution to the problem of missing data between the history
ending and the live data commencing.