Implemented QUERY_CANCELLED messages properly
Because a number of query types can be cancelled, the QUERY CANCELLED
message has been restructured to tell the unfortunate user as much
as it could about the original query. It is up to the caller to
decide if they want to try again or whatever.

Non-history query cancellations result in an exception which is
easily handled within the exporter code.