Queue config is no longer required if using export only mode
This was slightly more complicated than anticipated, as our main loop
was based around reading off the queue. If the queue doesn't exist, we
need to use something else to keep the exporter going while waiting
for client connections.

Therefore in export-only mode we now wait on the listen thread to
successfully join (which should only happen if something goes
horribly wrong). While waiting for the join, we also need to check
for SIGINT so we have to periodically timeout the join otherwise the
join will block and prevent us from getting the signal.