Catch OperationalError specifically when inserting into the database
This error usually indicates a postgres restart or some similar loss
of database connection. In that case, we don't want to try and
rollback (as we were previously doing) as this will fail miserably.

Instead, we catch the error, log it and return a new OPERATIONAL_ERROR
code. All primary parsers will attempt a reconnect in 5 seconds.

Also updated the rollback code to catch OperationalError and return
gracefully rather than crash NNTSC.