Attempt to handle OperationalError while querying data
If the database disappears during a DBWorker query, the worker
will now attempt to reconnect to the database. This reconnect is
subject to exponential backoff. Once connected, it will try the
failed query again.

If three disconnects occur without a query succeeding, the client
will be dropped. Most clients will probably just reconnect and
start over, but it will allow us to shed some connection load if
postgres all goes to hell.

Worker threads now have the ability to signal that the client
should be dropped. This will happen whenever a worker thread
exits -- in normal operation, a worker thread should only exit
if the client goes away or something went horribly wrong (e.g.
the completed job queue is full).

Most of this has not really been tested in the intended use case,
i.e. postgres restarts, as this is a little tricky to do without
breaking our production setups. Will schedule an outage soon for
some testing :)