Continue adding proper postgres restart handling
Database errors while updating the stream last timestamps are now dealt
with appropriately.

Added ability to distinguish between bad database names and an ongoing
restart when binding to a database. In the first case, we want to halt
the program but in the other we want to return an error code and retry
shortly.

Moved retry after an OperationalError into the generic insert functions
rather than trying to handle it in the dataparsers. We always want to
retry anyway.

Handle InvalidRequestError when trying to commit a transaction -- this
occurs when the database has been restarted and the current transaction
is invalid.