Change table structure to one-table-per-stream
Updated insertion code to create tables whenever a new stream is
registered and to insert new data into the correct table for each stream.

Updated all dataparsers (except HTTP) to use the new insert functions
and remove the old partitioning code. Testing on a fresh database shows
that this all seems to work as anticipated.