Queries are now performed one label at a time
This solves the problem where a stream could not only belong to more
than one label in a query. Previously, the duplicated stream would
only be included in the result for the first label that it matched
in our CASE statement.

It may also offer some performance improvements for "large" queries,
e.g. matrix data or graphs with many lines. This will need to be
tested further on large amounts of data to confirm.

Also, fixed another case where stream ids could be unnecessarily
duplicated in a part of a query. It's doubtful that this will help
performance (and shouldn't happen in the general case anyway) but
it helps keep things tidy and makes us less susceptible to stupid
bugs where we check the same thing multiple times.