Implemented new traceroute data table layout
This is a little different to before, in that we have two tables: one
for storing every path we've seen and another for each test result. The
path table saves us having to store duplicate paths for each similar
result, but makes insertion and querying a lot more complicated.

Extended lib/database.py to support adding a foreign key to a table, as
postgres does not copy foreign keys when using the LIKE syntax to create
a copy of an existing table template. This is necessary to maintain the
foreign key between path ids in our traceroute tables.

Also added a custom_insert function to lib/database.py. This allows us to
execute insertions that don't strictly qualify as "data", e.g. adding new
traceroute paths. The main differences when using custom_insert over
insert_data are 1. the inserted record is not automatically exported to
live clients and 2. the result is returned along with the error code. The
result is frequently useful in this case, e.g. returning the primary key of
the inserted row.

Added custom functions to amp_traceroute for sanitising query columns and
constructing the "union" clause of a query. These will be used in place
of the default approaches provided in lib/dbselect.py as they only really
work if there is a single data table. More in the next commit...