Separate traceroute paths table into one table per stream
As discussed earlier in the lab today, more per-stream separation
is probably better. This will hopefully make the query join much
faster in the general case because we only need to join with paths
that we know match the stream(s).

Also, there's more or less zero overlap across paths for different
streams, so there is no real gain in having a single table aside from
making the inserts and selects a little simpler.