Allow config options to be missing / commented out
If a config option is missing, NNTSC will try to use a sensible
default value instead, e.g. None for most of the database options.

LPI and RRD dataparsers will simply not run if required options are
missing but won't halt the rest of NNTSC.

Most live exporter options are still required but using -E will prevent
them being checked (which is the same as it was before).

Added log lines that explicitly describe the options being used by NNTSC.
This is necessary now as NNTSC may be setting default options for you and
it could get very confusing to have things fail due to default values
that you've never seen before. Obviously, password options are exempt
from being logged....