Added new 'smoke' and 'smokearray' aggregation functions to NNTSC
These functions replace the old 'percentile' method of generating
percentiles for plotting smokeping-style graphs.

'smoke' operates essentially the same as select_percentile_data()
did and is intended to be used in cases where there is one
measurement per row that the user wants to aggregate into
percentiles.

'smokearray' does a similar job but works on data where there
is an array of measurements in each row to be aggregated, i.e.
rrd-smokeping data.

Both aggregation functions are designed to be used with
select_aggregated_data, so select_percentile_data becomes
redundant at this point. It will eventually be deprecated.