Handle insert_stream being called for a stream that already exists
Instead of erroring and discarding the data, we now query for the
existing stream and return its stream id.

Re-organised flow of insert_stream to avoid creating entries in the
streams table or cloning data tables for the duplicate stream.

Removed foreign key constraint on stream_id for cloned data tables.