Make AMP dataparsers work with new database code
Added an optional 'cast' parameter to insert_data, which can be
used to specify CAST clauses on specific columns that require them.
For example, if cast = {'hop_rtt':'integer[]'}, then the insert
statement will contain CAST(hop_rtt as integer[]) rather than
just hop_rtt in the VALUES clause.

Removed a lot of commits from the new database code. This is mainly intended
to solve problems where some (but not all) of the steps involved
in creating a new stream are completed but then an error or interrupt
occurs. If the earlier steps are committed, we can end up with a stream
but no matching data or traceroute path table. Upon restarting NNTSC,
it will detect that the stream exists and will therefore try to insert
data into a table that may not exist.

Rename datacols in each AMP parser to be unique so as to avoid
namespace collisions.