Ensure all tables and stream rows are committed for new streams asap
The main issue here is that a new stream is exported to live clients
right away, who may then query for data for that new stream. If the
new data table isn't committed, we get programming errors because
the export thread can't query a table that doesn't actually exist
yet outside of a transaction in another process.

This has the undesirable side-effect of committing any inserted
data that had not yet been committed at the same time. Ideally going
forward, stream creation and data insertion probably should be handled
using different connections -- stream creation actions need to committed
right away but we want to hold back on committing data until we
are ready to acknowledge rabbit messages. Otherwise, we run the risk
of duplicating data if we hit an error between the new stream and
acknowledging the message.