Stop ProgrammingError from completely destroying the exporter
The error comes when someone is running an ampy for a long time
and the NNTSC database that it was pointing to has replaced its
streams table with a different one (i.e. running build_nntsc_db -F).

ampy has its own streams dict so its stream ids don't match what is
in the NNTSC database. ampy may start asking for data for streams
that don't exist anymore or are not part of the collection it thinks they
are, causing bad queries to be formed inside DBSelector.

The fix is very much a band-aid until the databaseinheritance branch
is finished and merged back. All this will do is catch the error, log it
and then pretend the query timed out so no graphs will be displayed.
Because the exception is caught, the NNTSC exporter process won't
crash and can continue to serve clients.

Long term, perhaps generate a unique key each time a database is generated
from scratch that clients can use to tell if they need to refresh their
local stream info?