Finish converting DatabaseCore and DBInsert to use new NNTSCCursor
Live exporting and database commits are now the responsibility of
the dataparser again. This allows dataparsers to ensure all stream-related
changes are committed before exporting new streams. It also allows them
to add fields to live data that is not being inserted into the primary data
table, e.g. the 'path' field for traceroute data.

Updated amp dataparsers to use the new API. LPI and RRD still to come.

DBSelect is also not yet using NNTSCCursor.