Process more AMP messages before acknowledging and commiting
This will increase insert performance for AMP when there is a
backlog of rabbit messages. It does mean there is a
tiny chance of the most recent data being delayed if messages
stop appearing in the queue for some reason but if this happens,
you've probably got much bigger problems.