Close cursors that have performed selects once we're done with the results
Best practice for psycopg2 is to close cursors after processing
the results, as the result will be cached and consume memory otherwise.

Closing dbselect data cursors should also help reduce memory usage on
the server.

Cursors used for insertion, e.g. the streams and data cursors for the
DBInsert class, are closed after the inserted data is committed.