Change handling of database disconnect while inserting streams
The OperationalError is now returned rather than re-attempting
the insert upon reconnect. The reason for this is that there may
have been data inserts that were lost due to the disconnect and
we need to re-attempt those. If we try again instead, we'll
eventually return success and the higher level dataparser
won't know that a disconnect occurred (and therefore there is data
that needs to be reprocessed).