Store first and last timestamps in memcache
This means that we no longer need to constantly update the
streams table in the database, which was proving costly for
our insert speed.

Timestamps are keyed by stream id. If no cache entry is present,
a query for max(timestamp) or min(timestamp) is performed on the
data table for the stream and the result is cached.

Whenever new data is inserted, the timestamp is also inserted in
the cache as the last timestamp. Whenever a new stream is created,
the timestamp is inserted in the first timestamp cache (except
in the case of RRD).

If a cache lookup is successful, the cache timeout is reset as we
want to avoid querying the data table wherever possible. We can
be certain that the cache entry will be updated if the last timestamp
changes, so not expiring frequently-used entries should be ok.