Removed the streams table from NNTSC
With first and last timestamp being memcached, the streams
table has no real reason to exist anymore. Updated all insertion
and querying code to no longer refer to the streams table.

Renamed 'label' in our data queries to 'nntsclabel' to reduce
the chance of it clashing with a column in a data table schema.