Change default stream cache time to 'forever'
Aside from memory usage, there's no real reason to timeout
the stream timestamps.  Timing them out just means we have to
query the inactive streams once a day, which is slow and annoying.

Only reset the cache timeout if we aren't caching forever, otherwise
this is just a waste of time.