Timestamp cache now operates at a collection level
Fetching from the cache for each individual stream was too
slow, especially when loading a matrix which can consist of
100s of streams.

Now, each collection caches a dictionary of stream ids to
(first, last) timestamp tuples. The dictionary is fetched once
per database query and the labels are filtered for active streams
before doing any queries (rather than filtering as part of the
query preparation process).

Any streams that need to be updated due to not being in cache
are updated individually rather than just writing the whole dict
back to cache when we're done querying. This is because the insert
process may have updated timestamps for streams while we were
querying and we don't want to overwrite those with previously
fetched values.

Changed default cache time to unlimited for stream timestamps. We need
a certain degree of permanence with our timestamps, especially since
querying the data tables to recalculate them is a slow operation.