Use exceptions rather than returning error codes
This means that error-handling is only generally needed in two places: where
the error originally occurs and at the point where the db operation was
called.

In practical terms, we can get rid of a crapload of:
    err = doSomething()
    if err != DB_NO_ERROR:
        return err

and just have:
    doSomething()

knowing that any exceptions will be travel up the stack to be caught
and handled by the original caller.

This does mean we've lost a bit of the more explicit logging when something
goes wrong at a particular step of a multi-stage operation, but I've done
my best to ensure that at least some logging happens when the error occurs
and hopefully that should make it clear exactly what operation failed,
e.g. which phase of creating a new stream caused the problem? Inserting the
stream in the streams table, cloning the data table, committing, freeing
the cursor? Hopefully psycopg2 will give us enough of an error message that
we can figure this out.

Removed unused return values from insert_stream and _generate_from in
DBInsert and DBSelector respectively.