Fix bug where no new collections could be added
build_nntsc_db always attempts to create the stream id sequence and
raises an exception if the sequence already exists. Due to recent
changes to error handling, this exception was causing build_nntsc_db
to bail on adding new collections, instead of catching it and ignoring
it.

Changed the exception reported in the duplicate sequence case so that it
is easier to distinguish from a genuine error.