Automatically export live data for new streams
Clients that wish to subscribe to a newly created stream
often miss out on the first data point -- the data is not
committed to the database when the client subscribe message
comes in so we don't return any history for the new stream and
the live data for that stream is normally discarded because no
clients have had a chance to subscribe to the stream yet.

This change automatically exports the first observed live data
for each new stream to all clients that have subscribed to a
stream for the same collection. This means that if the client does
wish to subscribe, they will get the otherwise missing first
data point. If they don't care about it, they should simply ignore
the data.