Try to touch streamcache a lot less often
Keeping the stream cache completely up to date was killing performance
on skeptic, so now timestamp updates are stored in memory and written
to the cache every 5 mins (instead of after every update).

Timestamps that cannot be found in cache at all are stored immediately
still, as the alternative is doing the database query again which is
even slower.