Fix bug in caching first timestamps
Any existing streams that weren't cached when NNTSC started
would get any subsequently fetched and cached timestamps overwritten
by the regular timestamp updates.

For last timestamp, this was generally ok since the timestamp from
inserted data would supercede the timestamp found by querying the
data for max(timestamp). However, for first timestamp the queried
timestamp would get overwritten with None, as update_timestamp would
only set the first timestamp if the stream was new.

Instead, first timestamps are only set following a query for
min(timestamp). If someone asks for the first timestamp, we'll query
for it and add it to the cache but we won't try to automatically
set it whenever a new stream is created, since we don't want to be
having to update the cache every time a new stream gets created.