Write updated timestamps based on realtime
Previously, we'd write the updated timestamps every 5 mins of
measurement time, which is fine if we're keeping up with incoming
measurements but creates a lot of unnecessary load if we're trying
to catch up on a large message queue.

Using realtime instead doesn't make us update the cache any less often
if we're keeping up and should allow us to get through a large backlog
a little bit faster, so seems to be win-win to me.