Include collection id with exported live data
This allows the exporter to avoid sending unwanted live data
to a client when the stream id is shared across multiple
collections, e.g. amp-astraceroute and amp-traceroute.

Fixed inconsistencies in variable naming in exporter.py

Ensure all relevant path information is retained when working out
the most common AS path, e.g. the path itself, unique AS counts etc,
as we may need to export this information to live clients.