Added unsubscribe support to exporter
Streams are registered as subscribed before passing the subscribe
job to the worker threads.

Streams are not filtered based on activity if the period being queried
is shorter than the frequency which we update the activity cache. This
fixes the bug where you could ask for 2 minutes of history and get
nothing, despite there being data available.