Don't fetch from RRD if no new data is expected
If the difference between the start and end times that we're asking
for is zero, we can probably get away with not trying to fetch from
the RRD file. Might save us a bit of file I/O, might not (if the python
RRD library is smart) but seems like a good idea regardless.