Trigger a retroactive continuous query for late data
If AMP data for a given bin arrives after the automatic continuous query,
that bin will otherwise never be updated to include the new measurement. Now
we try to estimate if a measurement is late and arrived outside of its bin
and force the continuous query to be re-run.

Testing so far: managed to go a whole weekend without any undersized 5 minute
bins, whereas previously 5-10% of bins were missing a measurement.