Always manually run the continuous queries ourselves
I don't trust influx to automatically run the query at the right
time, especially given how AMP results can be delayed, and trying to
work out when we should be running it manually wasn't working since
the upgrade to influx 0.11. In the end, since we were running the CQ
manually most of the time anyway, I've just given up on trying to figure
out if we can get away with NOT running it.