Fix binning issues caused by change in influx behaviour in 0.11
The timestamp returned when using the last() function is now the earliest
timestamp from the 'where' clause. In our case, this was always going to be
zero (because we had no where clause) so we were not correctly working out
when we had last aggregated.

Replaced the query with an "order by desc/asc limit 1" to avoid this problem.