Add 'requests' column to amp-dns data table
Since influx won't let us do aggregation on the timestamp column, we need a
new method of calculating the number of DNS tests run during a bin. I've added
a 'requests' column to the amp-dns data table, which is always set to 1 for a
raw DNS result and the CQs will sum to generate an aggregate count of DNS
tests run during that time period.

Also added a bit of hackery to ensure that ampy can work with both influx-nntsc
and postgres-nntsc without having to know which one it is talking to.