NNTSC can now query both influx and postgres dbs for required data
When transitioning from a postgres-based NNTSC to an influx-based one,
this will allow us to keep all of the old data in the postgres database
as NNTSC will query the postgres database if the time period queried for
dates back to before the database change.

Also fixed an annoying issue where aggregated data fetches would stop several
bins short of the current time because the continuous query has not been run
for that rollup yet -- usually because the machine is memory-starved. Now,
NNTSC will do its own aggregation for all bins after the last bin present in
the rollup table, not just the most recent bin.