Ensure all rows for the same label appear in sequence
My recent change tended to split rows for a given label into two groups,
one for postgres data and one for influx data. The exporter expects all
rows for the same label to appear in a single block, so this wasn't going
to work.

To compensate for losing one of Influx's main benefits -- the ability to
deal with all labels in a single query -- I've added some shortcut cases for
queries where we can be pretty certain there will be no postgres data. In
those cases, we can get away with doing a single query because we don't have
to worry about interleaving our results with the postgres data.