Periodically purge unused AS and IP paths
The traceroute parser keeps a record of previously observed paths in memory
so that we didn't need to keep querying the database to see if the path
already existed. Previously, we didn't do any pruning on the path list so
a long-running NNTSC could end up wasting a lot of memory storing paths that
are never likely to be seen again.

Instead, we now prune paths that have not been seen in the past hour. Pruning
happens hourly, so shouldn't be too onerous.