Fix bad aggregated data query results in influx
Bug was observed as old data being appended to aggregated data, causing
backwards lines to appear on graphs. The problem mostly arose if the
aggregation was way behind due to NNTSC trying to aggregate the extra
data prior to servicing the query, but the problem was that we were going
back a bit too far (i.e. ignoring the requested start time and always going
back to the time we last aggregated).